% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampledata}
\alias{exampledata}
\title{Example dataset for lefser}
\format{
An ExpressionSet with 1585 features, 199 samples
\describe{
  \item{study_condition}{adenoma, control, CRC}
  \item{age_category}{adult, senoir}
  ...
}
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/25432777/}
}
\usage{
exampledata
}
\description{
The ZellerG_2014 dataset contains microbiome count data for CRC patients and controls.
It was for curatedMetagenomicData using the script in the package directory "data/raw".
}
\keyword{datasets}
