% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lefsePlotClad.R
\name{lefsePlotClad}
\alias{lefsePlotClad}
\title{Plot Cladogram}
\usage{
lefsePlotClad(x)
}
\arguments{
\item{x}{An object of class "lefesr_df_all".}
}
\value{
A ggtree object.
}
\description{
Plot Cladogram
}
