% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lefsePlotFeat.R
\name{lefsePlotFeat}
\alias{lefsePlotFeat}
\title{Plot Feature}
\usage{
lefsePlotFeat(res, fName)
}
\arguments{
\item{res}{An object of class lefser_df,
output of the \code{lefser} function.}

\item{fName}{A character string. The name of a feature in the lefser_df
object.}
}
\value{
A ggplot object.
}
\description{
\code{lefsePlotFeat} plots the abundance data of a DA feature across all
samples.
}
\details{
The solid lines represent the mean by group or by group+block
(if the block variable is present).
The dashed lines represent the median by group or by group+block
(if the block variable is present).
}
\examples{

data(zeller14)
zeller14 <- zeller14[, zeller14$study_condition != "adenoma"]
tn <- get_terminal_nodes(rownames(zeller14))
zeller14tn <- zeller14[tn,]
zeller14tn_ra <- relativeAb(zeller14tn)

# (1) Using classes only
res_group <- lefser(zeller14tn_ra,
                    groupCol = "study_condition")
# (2) Using classes and sub-classes
res_block <- lefser(zeller14tn_ra,
                    groupCol = "study_condition",
                    blockCol = "age_category")
plot_group <- lefsePlotFeat(res_group, res_group$features[[1]])
plot_block <- lefsePlotFeat(res_block, res_block$features[[2]])

}
