% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lefserPlotClad.R
\name{lefserAllRanks}
\alias{lefserAllRanks}
\title{Run lefser on all taxonomic levels}
\usage{
lefserAllRanks(relab, ...)
}
\arguments{
\item{relab}{A SummarizedExperiment.}

\item{...}{Arguments passed to the \code{lefser} function.}
}
\value{
An object of class 'lefser_df_all' and 'data.frame'.
}
\description{
Run lefser on all taxonomic levels
}
\examples{

data("zeller14")
z14 <- zeller14[, zeller14$study_condition != "adenoma"]
tn <- get_terminal_nodes(rownames(z14))
z14tn <- z14[tn, ]
z14tn_ra <- relativeAb(z14tn)

resAll <- lefserAllRanks(relab = z14tn_ra, groupCol = "study_condition")

}
