% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lefser.R
\name{lefser}
\alias{lefser}
\title{R implementation of the LEfSe method}
\usage{
lefser(
  expr,
  p.threshold = 0.05,
  lda.threshold = 2,
  groupCol = "GROUP",
  blockCol = NULL,
  assay = 1L
)
}
\arguments{
\item{expr}{A \code{\linkS4class{SummarizedExperiment}} with expression data.}

\item{p.threshold}{numeric(1) The p-value for the Kruskal-Wallis Rank Sum
Test. If 'blockCol' is provided, it indicates the p-value for the Wilcoxon
Rank-Sum Test (default 0.05).}

\item{lda.threshold}{The effect size threshold.
The default is at 2.0.}

\item{groupCol}{character(1) Column name in `colData(expr)` indicating
groups, usually a factor with two levels (e.g., `c("cases", "controls")`;
default "GROUP").}

\item{blockCol}{character(1) Column name in `colData(expr)` indicating the
blocks, usually a factor with two levels (e.g., `c("adult", "senior")`;
default NULL).}

\item{assay}{The i-th assay matrix in the `SummarizedExperiment` ('expr').
Defaults to 1.}
}
\value{
The function returns a dataframe with two columns, which are
names of microorganisms and their LDA scores.
}
\description{
Perform a LEfSe analysis: the function carries out differential analysis
between two sample groups for multiple microorganisms and uses linear discirminant analysis
to establish their effect sizes. Subclass information for each class can be incorporated
into the analysis (see examples). Microorganisms with large differences between two sample groups
are identified as biomarkers.
}
\examples{
    # (1) Using classes only
    data(zeller14)
    # exclude 'adenoma'
    zeller14 <- zeller14[, zeller14$study_condition != "adenoma"]
    # assign '0' class to 'conrol' and '1' to 'CRC' (i.e., colorectal cancer)
    zeller14$GROUP <- ifelse(zeller14$study_condition == "control", 0, 1)
    results <- lefser(zeller14)
    head(results)

    # (2) Using classes and sublasses
    data(zeller14)
    # exclude 'adenoma'
    zeller14 <- zeller14[, zeller14$study_condition != "adenoma"]
    # assign '0' class to 'conrol' and '1' to 'CRC' (i.e., colorectal cancer)
    zeller14$GROUP <- ifelse(zeller14$study_condition == "control", 0, 1)
    # assign '0' class to 'adult' and '1' to 'senior'
    zeller14$BLOCK <- ifelse(zeller14$age_category == "adult", 0, 1)
    results <- lefser(zeller14)
    head(results)
}
