% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lefserPlotClad.R
\name{lefserAllRanks}
\alias{lefserAllRanks}
\title{Run lefser on all taxonomic levels}
\usage{
lefserAllRanks(relab, ...)
}
\arguments{
\item{relab}{A SummarizedExperiment.}

\item{...}{Arguments passed to the \code{lefser} function.}
}
\value{
An object of class 'lefser_df_all' and 'data.frame'.
}
\description{
Run lefser on all taxonomic levels
}
