% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lefserAnalysis.R
\name{lefserAnalysis}
\alias{lefserAnalysis}
\title{Title
Perform a LEfSe analysis}
\usage{
lefserAnalysis(expr)
}
\arguments{
\item{expr}{Explain expr}
}
\value{
Explain what the function returns
}
\description{
Title
Perform a LEfSe analysis
}
