% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lefserAnalysis.R
\name{lefserAnalysis}
\alias{lefserAnalysis}
\title{Title
Perform a LEfSe analysis}
\usage{
lefserAnalysis(expr, kw.threshold = 0.05, wilcoxon.threshold = 0.05)
}
\arguments{
\item{expr}{Explain expr}

\item{wilcoxon.threshold}{}
}
\value{
Explain what the function returns
}
\description{
Title
Perform a LEfSe analysis
}
\examples{
data(exampledata)
exampledata <- exampledata[, exampledata$study_condition != "adenoma"]
exampledata$GROUP <- ifelse(exampledata$study_condition == "control", 0, 1)
# keep only taxa with non-zero counts in at least 10 specimens:
results <- lefserAnalysis(exampledata)
head(results)
}
