% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lefserPlot.R
\name{lefserPlot}
\alias{lefserPlot}
\title{Title add title}
\usage{
lefserPlot(scores_df)
}
\arguments{
\item{scores_df}{explain scores_df}
}
\value{
explain the return value
}
\description{
Title add title
}
\examples{
example("lefseranalysis")
lefserPlot(results)
}
