% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lefserPlotClad.R
\name{lefserPlotClad}
\alias{lefserPlotClad}
\title{LEfSer plot cladogram}
\usage{
lefserPlotClad(df, colors = "c", showTipLabels = FALSE, showNodeLabels = "p")
}
\arguments{
\item{df}{An object of class "lefser_df" or "lefesr_df_all".}

\item{colors}{dfsdf}

\item{showTipLabels}{sdfsd}

\item{showNodeLabels}{sdfsdfs}
}
\value{
A ggtree object.
}
\description{
LEfSer plot cladogram
}
