% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lefserPlotClad.R
\name{lefserPlotClad}
\alias{lefserPlotClad}
\title{Plot Cladogram}
\usage{
lefserPlotClad(x)
}
\arguments{
\item{x}{An object of class "lefser_df" or "lefesr_df_all".}
}
\value{
A ggtree object.
}
\description{
Plot Cladogram
}
