% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zeller14}
\alias{zeller14}
\title{Example dataset for lefser}
\format{
A \link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}
with 1585 features, 199 samples
\describe{
\item{study_condition}{adenoma, control, CRC}
\item{age_category}{adult, senior}
}
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/25432777/}
}
\usage{
data("zeller14")
}
\description{
The ZellerG_2014 dataset contains microbiome count data for CRC patients and
controls. It was for curatedMetagenomicData using the script in the package
directory "data-raw".
}
\keyword{datasets}
